feat(ms-teams): Added MSTeamsJobDetailsAccessor interface.
