+++ alert-common/src/main/java/com/synopsys/integration/alert/common/persistence/accessor/MSTeamsJobDetailsAccessor.java
@@ -1,5 +1,5 @@
 /*
- * alert-database
+ * alert-common
  *
  * Copyright (c) 2021 Synopsys, Inc.
  *
@@ -20,27 +20,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.database.job.msteams;
-
-import java.util.UUID;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+package com.synopsys.integration.alert.common.persistence.accessor;
 
 import com.synopsys.integration.alert.common.persistence.model.job.details.MSTeamsJobDetailsModel;
 
-@Component
-public class MSTeamsJobDetailsAccessor {
-    private final MSTeamsJobDetailsRepository msTeamsJobDetailsRepository;
-
-    @Autowired
-    public MSTeamsJobDetailsAccessor(MSTeamsJobDetailsRepository msTeamsJobDetailsRepository) {
-        this.msTeamsJobDetailsRepository = msTeamsJobDetailsRepository;
-    }
-
-    public MSTeamsJobDetailsEntity saveMSTeamsJobDetails(UUID jobId, MSTeamsJobDetailsModel msTeamsJobDetails) {
-        MSTeamsJobDetailsEntity jobDetailsToSave = new MSTeamsJobDetailsEntity(jobId, msTeamsJobDetails.getWebhook());
-        return msTeamsJobDetailsRepository.save(jobDetailsToSave);
-    }
-
+public interface MSTeamsJobDetailsAccessor extends JobDetailsAccessor<MSTeamsJobDetailsModel> {
 }
