+++ alert-database/src/main/java/com/synopsys/integration/alert/database/job/msteams/DefaultMSTeamsJobDetailsAccessor.java
@@ -1,0 +1,53 @@
+/*
+ * alert-database
+ *
+ * Copyright (c) 2021 Synopsys, Inc.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.synopsys.integration.alert.database.job.msteams;
+
+import java.util.Optional;
+import java.util.UUID;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import com.synopsys.integration.alert.common.persistence.accessor.MSTeamsJobDetailsAccessor;
+import com.synopsys.integration.alert.common.persistence.model.job.details.MSTeamsJobDetailsModel;
+
+@Component
+public class DefaultMSTeamsJobDetailsAccessor implements MSTeamsJobDetailsAccessor {
+    private final MSTeamsJobDetailsRepository msTeamsJobDetailsRepository;
+
+    @Autowired
+    public DefaultMSTeamsJobDetailsAccessor(MSTeamsJobDetailsRepository msTeamsJobDetailsRepository) {
+        this.msTeamsJobDetailsRepository = msTeamsJobDetailsRepository;
+    }
+
+    public MSTeamsJobDetailsEntity saveMSTeamsJobDetails(UUID jobId, MSTeamsJobDetailsModel msTeamsJobDetails) {
+        MSTeamsJobDetailsEntity jobDetailsToSave = new MSTeamsJobDetailsEntity(jobId, msTeamsJobDetails.getWebhook());
+        return msTeamsJobDetailsRepository.save(jobDetailsToSave);
+    }
+
+    @Override
+    public Optional<MSTeamsJobDetailsModel> retrieveDetails(UUID jobId) {
+        return msTeamsJobDetailsRepository.findById(jobId)
+                   .map(entity -> new MSTeamsJobDetailsModel(entity.getJobId(), entity.getWebhook()));
+    }
+}
